Move tests to a separate `#[cfg(test)]` module
